+++ app/Http/Requests/ExpenseRequest.php
@@ -2,7 +2,11 @@
 
 namespace App\Http\Requests;
 
+use App\Enums\ExpenseCategories;
+use Illuminate\Contracts\Validation\Validator;
 use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Http\Exceptions\HttpResponseException;
+use Illuminate\Validation\Rules\Enum;
 
 class ExpenseRequest extends FormRequest
 {
@@ -23,12 +27,22 @@ class ExpenseRequest extends FormRequest
     {
         return [
             'description'    => 'required|string',
-            'category'       => 'required',
+            'category'       => ['required', new Enum(ExpenseCategories::class)],
             'amount'         => 'required|numeric',
             'currency'       => 'required|string',
             'expense_date'   => 'required|date',
             'note'          => 'nullable|string',
             'recurring_type' => 'nullable|string|in:once,monthly',
         ];
+    }
+    public function failedValidation(Validator $validator)
+    {
+        throw new HttpResponseException(
+            response()->json([
+                'success' => false,
+                'message' => $validator->errors()->first(),
+                'errors' => $validator->errors()
+            ], 422)
+        );
     }
 }