+++ database/migrations/2026_03_11_032809_create_goals_table.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Enums\GoalCategories;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -13,6 +14,14 @@ return new class extends Migration
     {
         Schema::create('goals', function (Blueprint $table) {
             $table->id();
+            $table->foreignId('user_id')->constrained()->cascadeOnDelete();
+            $table->string('title');
+            $table->enum('category', array_column(GoalCategories::cases(), 'value'))->default(GoalCategories::Other->value);
+            $table->date('deadline_date');
+            $table->time('deadline_time');
+            $table->text('note')->nullable();
+            $table->boolean('is_completed')->default(false);
+            $table->timestamp('completion_date')->nullable();
             $table->timestamps();
         });
     }